+++ manta-crypto/src/merkle_tree/tree.rs
@@ -26,8 +26,8 @@
 
 use crate::{
     accumulator::{
-        self, Accumulator, ConstantCapacityAccumulator, ExactSizeAccumulator, MembershipProof,
-        OptimizedAccumulator,
+        self, Accumulator, BatchInsertion, ConstantCapacityAccumulator, ExactSizeAccumulator,
+        MembershipProof, OptimizedAccumulator,
     },
     eclair::{
         self,
@@ -41,6 +41,7 @@ use crate::{
         path::{constraint::PathVar, CurrentPath, Path},
     },
 };
+use alloc::vec::Vec;
 use core::{fmt::Debug, hash::Hash, marker::PhantomData};
 use manta_util::{
     codec::{Decode, DecodeError, Encode, Read, Write},
@@ -301,6 +302,64 @@ where
         self.push_digest(parameters, || parameters.digest(leaf))
     }
 
+    /// Inserts `leaf_digests` in `self`, see [`maybe_push_digest`] for more details.
+    /// Returns `true` if all the insertions succeed and `false` otherwise.
+    ///
+    /// # Implementation Note
+    ///
+    /// By default, this method calls [`maybe_push_digest`] individually for each leaf digest
+    /// in `leaf_digests` till failure. Custom implementations of this method to improve performance
+    /// must return the same result as the default implementation.
+    ///
+    /// [`maybe_push_digest`]: Tree::maybe_push_digest
+    #[inline]
+    fn batch_maybe_push_digest<F>(
+        &mut self,
+        parameters: &Parameters<C>,
+        leaf_digests: F,
+    ) -> Option<bool>
+    where
+        F: FnOnce() -> Vec<LeafDigest<C>>,
+    {
+        let leaf_digests = leaf_digests();
+        if leaf_digests.is_empty() {
+            return None;
+        }
+        for leaf_digest in leaf_digests {
+            if !self.maybe_push_digest(parameters, || Some(leaf_digest))? {
+                return Some(false);
+            }
+        }
+        Some(true)
+    }
+
+    /// Inserts `leaf_digests` in `self`. Returns `true` if all the insertions
+    /// succeed or if there were no `leaf_digests` and `false` otherwise.
+    #[inline]
+    fn batch_push_digest<F>(&mut self, parameters: &Parameters<C>, leaf_digests: F) -> bool
+    where
+        F: FnOnce() -> Vec<LeafDigest<C>>,
+    {
+        self.batch_maybe_push_digest(parameters, leaf_digests)
+            .unwrap_or(true)
+    }
+
+    /// Inserts the digests of `leaves` at the next available leaf node of the tree, returning
+    /// `false` if the leaves could not be inserted because the tree has exhausted its capacity.
+    #[inline]
+    fn batch_push<'a, I>(&mut self, parameters: &Parameters<C>, leaves: I) -> bool
+    where
+        Leaf<C>: 'a,
+        I: IntoIterator<Item = &'a Leaf<C>>,
+    {
+        self.batch_push_digest(parameters, || {
+            leaves
+                .into_iter()
+                .map(|leaf| parameters.digest(leaf))
+                .collect()
+        })
+    }
+
     /// Appends an iterator of leaf digests at the end of the tree, returning the iterator back
     /// if it could not be inserted because the tree has exhausted its capacity.
     ///
@@ -449,6 +508,64 @@ where
     #[inline]
     fn push_provable(&mut self, parameters: &Parameters<C>, leaf: &Leaf<C>) -> bool {
         self.push_provable_digest(parameters, move || parameters.digest(leaf))
+    }
+
+    /// Inserts `leaf_digests` provably in `self`, see [`maybe_push_provable_digest`] for more details.
+    /// Returns `true` if all the insertions succeed and `false` otherwise.
+    ///
+    /// # Implementation Note
+    ///
+    /// By default, this method calls [`maybe_push_provable_digest`] individually for each leaf digest
+    /// in `leaf_digests` till failure. Custom implementations of this method to improve performance
+    /// must return the same result as the default implementation.
+    ///
+    /// [`maybe_push_provable_digest`]: WithProofs::maybe_push_provable_digest
+    #[inline]
+    fn batch_maybe_push_provable_digest<F>(
+        &mut self,
+        parameters: &Parameters<C>,
+        leaf_digests: F,
+    ) -> Option<bool>
+    where
+        F: FnOnce() -> Vec<LeafDigest<C>>,
+    {
+        let leaf_digests = leaf_digests();
+        if leaf_digests.is_empty() {
+            return None;
+        }
+        for leaf_digest in leaf_digests {
+            if !self.maybe_push_provable_digest(parameters, || Some(leaf_digest))? {
+                return Some(false);
+            }
+        }
+        Some(true)
+    }
+
+    /// Inserts `leaf_digests` provably in `self`. Returns `true` if all the insertions
+    /// succeed or if there were no `leaf_digests` and `false` otherwise.
+    #[inline]
+    fn batch_push_provable_digest<F>(&mut self, parameters: &Parameters<C>, leaf_digests: F) -> bool
+    where
+        F: FnOnce() -> Vec<LeafDigest<C>>,
+    {
+        self.batch_maybe_push_provable_digest(parameters, leaf_digests)
+            .unwrap_or(true)
+    }
+
+    /// Inserts the digests of `leaves` provably at the next available leaf node of the tree, returning
+    /// `false` if the leaves could not be inserted because the tree has exhausted its capacity.
+    #[inline]
+    fn batch_push_provable<'a, I>(&mut self, parameters: &Parameters<C>, leaves: I) -> bool
+    where
+        Leaf<C>: 'a,
+        I: IntoIterator<Item = &'a Leaf<C>>,
+    {
+        self.batch_push_provable_digest(parameters, || {
+            leaves
+                .into_iter()
+                .map(|leaf| parameters.digest(leaf))
+                .collect()
+        })
     }
 
     /// Returns the path for the leaf stored at the given `index` if it exists.
@@ -913,6 +1030,19 @@ where
         self.tree.push(&self.parameters, leaf)
     }
 
+    /// Inserts `leaves` at the next avaiable leaf node of the tree, returning `true` if all
+    /// leaf insertions succeed and `false` otherwise.
+    ///
+    /// See [`Tree::batch_push`] for more.
+    #[inline]
+    pub fn batch_push<'a, I>(&mut self, leaves: I) -> bool
+    where
+        Leaf<C>: 'a,
+        I: IntoIterator<Item = &'a Leaf<C>>,
+    {
+        self.tree.batch_push(&self.parameters, leaves)
+    }
+
     /// Appends an iterator of leaves at the end of the tree, returning `false` if the `leaves`
     /// could not be inserted because the tree has exhausted its capacity.
     ///
@@ -993,6 +1123,20 @@ where
         T: WithProofs<C>,
     {
         self.tree.push_provable(&self.parameters, leaf)
+    }
+
+    /// Inserts `leaves` provably at the next avaiable leaf node of the tree, returning `true` if all
+    /// provable leaf insertions succeed and `false` otherwise.
+    ///
+    /// See [`WithProofs::batch_push_provable`] for more.
+    #[inline]
+    pub fn batch_push_provable<'a, I>(&mut self, leaves: I) -> bool
+    where
+        T: WithProofs<C>,
+        Leaf<C>: 'a,
+        I: IntoIterator<Item = &'a Leaf<C>>,
+    {
+        self.tree.batch_push_provable(&self.parameters, leaves)
     }
 
     /// Returns the path for the leaf stored at the given `index` if it exists.
@@ -1150,6 +1294,32 @@ where
             .position(&self.parameters.digest(item))
             .map(move |i| self.tree.remove_path(i))
             .unwrap_or(false)
+    }
+}
+
+impl<C, T> BatchInsertion for MerkleTree<C, T>
+where
+    C: Configuration + ?Sized,
+    T: Tree<C> + WithProofs<C>,
+    InnerDigest<C>: Clone + PartialEq,
+    Parameters<C>: Clone,
+{
+    #[inline]
+    fn batch_insert<'a, I>(&mut self, items: I) -> bool
+    where
+        Self::Item: 'a,
+        I: IntoIterator<Item = &'a Self::Item>,
+    {
+        self.batch_push_provable(items)
+    }
+
+    #[inline]
+    fn batch_insert_nonprovable<'a, I>(&mut self, items: I) -> bool
+    where
+        Self::Item: 'a,
+        I: IntoIterator<Item = &'a Self::Item>,
+    {
+        self.batch_push(items)
     }
 }